refactor: add StepThree, StepFour, and StepFive components
